requirements.txt
requests
Pillow
torch
torchvision
numpy
tqdm
matplotlib
wandb
fastapi
uvicorn
python-multipart